refactor: RedisService에서 Lock 구현
