+++ src/main/java/gift/service/ProductService.java
@@ -1,6 +1,5 @@
 package gift.service;
 
-import gift.common.annotation.RedissonLock;
 import gift.common.exception.EntityNotFoundException;
 import gift.controller.dto.request.OptionRequest;
 import gift.controller.dto.response.OptionResponse;
@@ -120,7 +119,7 @@ public class ProductService {
         return OptionResponse.from(option);
     }
 
-    @RedissonLock(value = "#id + ':' + #optionId")
+    @Transactional
     public int subtractQuantity(Long id, Long optionId, int amount) {
         Product product = productRepository.findProductAndOptionByIdFetchJoin(id)
                 .orElseThrow(() -> new EntityNotFoundException("Product with id " + id + " not found"));
